+++ tests/test_async_master.py
@@ -1,4 +1,5 @@
-# Copyright (c) 2014-2018, XMOS Ltd, All rights reserved
+# Copyright 2014-2021 XMOS LIMITED.
+# This Software is subject to the terms of the XMOS Public Licence: Version 1.
 import xmostest
 from i2c_master_checker import I2CMasterChecker
 import os
@@ -39,7 +40,7 @@ def do_master_test(arch, speed, impl, stop):
                               tester = tester)
 
 def runtest():
-  for arch in ['xs1', 'xs2']:
+  for arch in ['xs1', 'xs2', 'xcoreai']:
     for stop in ['stop', 'no_stop']:
       do_master_test(arch, 400, 'non_comb', stop)
       for impl in ['comb', 'non_comb']: