+++ tests/test_reg_ops_nack.py
@@ -1,4 +1,5 @@
-# Copyright (c) 2014-2018, XMOS Ltd, All rights reserved
+# Copyright 2014-2021 XMOS LIMITED.
+# This Software is subject to the terms of the XMOS Public Licence: Version 1.
 import xmostest
 from i2c_master_checker import I2CMasterChecker
 import os
@@ -39,5 +40,5 @@ def do_test(arch):
 
 def runtest():
   # See BUG 17936 - the xs1 is not fast enough to run this test
-  for arch in ['xs2']:
+  for arch in ['xs2', 'xcoreai']:
     do_test(arch)